updates to email
