+++ src/main/java/org/launchcode/demo/models/email/EmailServiceInterface.java
@@ -1,0 +1,8 @@
+//package org.launchcode.demo.models.email;
+//
+//
+//public interface EmailService {
+//
+//    boolean sendMail(EmailDetails details);
+//
+//}
