+++ src/main/resources/templates/user/email.html
@@ -25,7 +25,7 @@
     <input type="hidden" name="requestingUsername" th:value="${requestingUsername}">
 
     <input type="submit" class="btn btn-primary" value="Email Owner" />
-<!--    <p th:text="${success}"></p>-->
+
     <p><a href="../search">Continue Searching for Books</a></p>
     <p><a href="../user/index">Return to Profile Page</a></p>
