+++ src/main/resources/templates/user/emailsuccess.html
@@ -10,6 +10,7 @@
 <body>
 
 <h4 th:text="${success}"></h4>
+
 <p><a href="../search">Continue Searching for Books</a></p>
 <p><a href="../user/index">Return to Profile Page</a></p>
