+++ src/main/resources/templates/user/register.html
@@ -26,11 +26,11 @@
     </div>
     <div>Location
         <div>
-            <input type="radio" id="kansascity" value="0" th:field="${registerFormDTO.location}">
+            <input type="radio" id="kansascity" value="1" th:field="${registerFormDTO.location}">
             <label for="kansascity">Kansas City</label><br>
-            <input type="radio" id="philadelphia" value="1" th:field="${registerFormDTO.location}">
+            <input type="radio" id="philadelphia" value="2" th:field="${registerFormDTO.location}">
             <label for="philadelphia">Philadelphia</label><br>
-            <input type="radio" id="saintlouis" value="2" th:field="${registerFormDTO.location}">
+            <input type="radio" id="saintlouis" value="3" th:field="${registerFormDTO.location}">
             <label for="saintlouis">Saint Louis</label>
         </div>
     </div>
